Refactor path structure
